Revert "fix(compiler-cli): avoid broken references in .d.ts files due to @internal markers"
This reverts commit 69821fcef67f76cbfd9c3aa6db344c7f292c7d06.
