+++ packages/compiler-cli/src/ngtsc/diagnostics/src/error.ts
@@ -8,8 +8,7 @@
 
 import ts from 'typescript';
 
-import {ErrorCode} from './error_code';
-import {ngErrorCode} from './util';
+import {ErrorCode, ngErrorCode} from './error_code';
 
 export class FatalDiagnosticError {
   constructor(
