+++ packages/compiler-cli/src/ngtsc/diagnostics/src/error_code.ts
@@ -226,3 +226,33 @@ export enum ErrorCode {
    */
   SUGGEST_SUBOPTIMAL_TYPE_INFERENCE = 10002,
 }
+
+/**
+ * @internal
+ * Base URL for the error details page.
+ * Keep this value in sync with a similar const in
+ * `packages/core/src/render3/error_code.ts`.
+ */
+export const ERROR_DETAILS_PAGE_BASE_URL = 'https://angular.io/errors';
+
+/**
+ * @internal
+ * Contains a set of error messages that have detailed guides at angular.io.
+ * Full list of available error guides can be found at https://angular.io/errors
+ */
+export const COMPILER_ERRORS_WITH_GUIDES = new Set([
+  ErrorCode.DECORATOR_ARG_NOT_LITERAL,
+  ErrorCode.IMPORT_CYCLE_DETECTED,
+  ErrorCode.PARAM_MISSING_TOKEN,
+  ErrorCode.SCHEMA_INVALID_ELEMENT,
+  ErrorCode.SCHEMA_INVALID_ATTRIBUTE,
+  ErrorCode.MISSING_REFERENCE_TARGET,
+  ErrorCode.COMPONENT_INVALID_SHADOW_DOM_SELECTOR,
+]);
+
+/**
+ * @internal
+ */
+export function ngErrorCode(code: ErrorCode): number {
+  return parseInt('-99' + code);
+}
